#1 - Spring Data Rest 테스트 제외 처리
이 프로젝트의 비즈니스 로직으로 구현한 내용이 아니라 data rest 기능이고,
통합테스트라 무거우며 db에도 영향을 준다.
공부 목적으로도 의미는 있기 때문에 삭제하지는 않고
제외 처리해서 테스트 중에 실행되지 않게끔 처리
