+++ src/main/java/com/fastcampus/projectboard/config/JpaConfig.java
@@ -15,5 +15,4 @@ public class JpaConfig {
     public AuditorAware<String> auditorAware(){
         return () -> Optional.of("gps"); // TODO: 스프링 시큐리티로 인증 기능을 붙이게 될 때, 수정하자
     }
-
 }
