+++ src/main/resources/application.yaml
@@ -26,7 +26,3 @@ spring:
   data.rest:
     base-path: /api
     detection-strategy: annotated
-
-
-
-
