--- src/test/java/com/fastcampus/projectboard/DataRestTest.java
@@ -1,40 +1,0 @@
-package com.fastcampus.projectboard;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@DisplayName("Data Rest - API테스트")
-@Transactional
-@AutoConfigureMockMvc
-@SpringBootTest
-public class DataRestTest {
-
-    private MockMvc mvc;
-
-    public DataRestTest(@Autowired MockMvc mvc) {
-        this.mvc = mvc;
-    }
-
-    @DisplayName("[api] 게시글 리스트 조회")
-    @Test
-    void givenNothing_whenRequestingArticles_thenReturnsArticlesJsonResponse() throws Exception {
-        // Given
-
-        // When & Then
-        mvc.perform(get("/api/articles"))
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.valueOf("application/hal+json")));
-
-    }
-}
